Semaphore implementation start
Some real coding for implementing semaphore service with the prettyOS.
Currently, OS_SemCreate(), OS_SemPend() is implemented without any test.

OS_EVENT structure will be consider to hold the data of the any event object
such as semaphore or mailbox or ... etc.

The events that a task is relying on will be placed in a decreasing sorted queue
based on priority value to be easily to get the highest priority waiting task to respond
to event occurrence.

Signed-off-by: Yahia Farghaly <yahiafarghaly@gmail.com>