Added O(1) C implementation of clz instruction.
This is done using a table of fixed known values of leading zeros of values ranging from 0 to 255
plus some bitwise operations.

It consumes more CPU cycles than if a clz assembly instruction is provided but it's better than using
the known loop methods to count the number of leading zero bits in a given number.

Signed-off-by: Yahia Farghaly <yahiafarghaly@gmail.com>